# Work_Assignment/AppServer/src/IO.Swagger/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "PlugGoBeyond",
    "description" : "Rest API of the mobile app PlugGoBeyond of the project course LI IV in the program Mestrado Integrado em Engenharia Informática",
    "termsOfService" : "http://swagger.io/terms",
    "contact" : {
      "email" : "a34900/a82888/a85731/a86618/a89982@alunos.uminho.pt"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.0"
  },
...
  "servers" : [ {
    "url" : "/"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/vehicles" : {
      "get" : {
        "tags" : [ "vehicles" ],
        "summary" : "List all vehicles.",
        "operationId" : "findVehicles",
        "responses" : {
          "200" : {
            "description" : "Success.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Vehicle"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Operation error."
          }
        }
      },
      "post" : {
        "tags" : [ "vehicles" ],
        "summary" : "Add a vehicle into the system.",
        "operationId" : "addVehicle",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/VehicleInfo"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Success."
          },
          "400" : {
            "description" : "Invalid Parameter."
          }
        }
      }
    },
    "/vehicles/filterByIds" : {
      "get" : {
        "tags" : [ "vehicles" ],
        "summary" : "Find all trips by the ids given.",
        "description" : "Multiple vehicles can be requested to be given. It's only required that they have a comma token to separate them.",
        "operationId" : "findVehiclesByIds",
        "parameters" : [ {
          "name" : "ids",
          "in" : "query",
          "description" : "Ids of the vehicles.",
          "required" : true,
          "style" : "form",
          "explode" : false,
          "schema" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Vehicle"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid ids."
          }
        }
      }
    },
    "/vehicles/{registrationNr}" : {
      "get" : {
        "tags" : [ "vehicles" ],
        "summary" : "Find a vehicle by it's registration number.",
        "description" : "Returns the vehicle",
        "operationId" : "getVehicleByRegsNr",
        "parameters" : [ {
          "name" : "registrationNr",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Vehicle"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid registration number."
          },
          "404" : {
            "description" : "Vehicle not found."
          }
        }
      },
      "put" : {
        "tags" : [ "vehicles" ],
        "summary" : "Updates a vehicle.",
        "operationId" : "updateVehicle",
        "parameters" : [ {
          "name" : "registrationNr",
          "in" : "path",
          "description" : "Registration number of the vehicle that needs to be updated.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Vehicle"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "400" : {
            "description" : "Invalid registration number."
          },
          "404" : {
            "description" : "Vehicle not found."
          }
        }
      },
      "delete" : {
        "tags" : [ "vehicles" ],
        "summary" : "Removes a vehicle.",
        "operationId" : "deleteVehicle",
        "parameters" : [ {
          "name" : "username",
          "in" : "query",
          "description" : "Username of the user that want to delete the vehicle.",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "registrationNr",
          "in" : "path",
          "description" : "Registration number of the vehicle that needs to be removed.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "400" : {
            "description" : "Invalid registration number."
          },
          "404" : {
            "description" : "Vehicle not found."
          }
        }
      }
    },
    "/vehicles/{registrationNr}/getTypeCode" : {
      "get" : {
        "tags" : [ "vehicles" ],
        "summary" : "Find a vehicle's type code by it's registration number.",
        "description" : "Returns the vehicle's type code.",
        "operationId" : "getVehicleTypeCode",
        "parameters" : [ {
          "name" : "registrationNr",
          "in" : "path",
          "description" : "Registration number of the vehicle that we want to extract type code.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Success.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string",
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid registration number."
          },
          "404" : {
            "description" : "Vehicle not found."
          }
        }
      }
    },
    "/vehicles/{registrationNr}/putConsumption" : {
      "put" : {
        "tags" : [ "vehicles" ],
        "summary" : "Updates the last consumption of a vehicle.",
        "description" : "Returns success if the operation is completed.",
        "operationId" : "updateLastConsumption",
        "parameters" : [ {
          "name" : "registrationNr",
          "in" : "path",
          "description" : "Registration number of the vehicle that needs to be updated.",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateConsumption"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Success."
          },
          "400" : {
            "description" : "Invalid registration number."
          },
          "404" : {
            "description" : "Vehicle not found."
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Vehicle" : {
        "type" : "object",
        "properties" : {
          "registrationNumber" : {
            "type" : "string"
          },
          "typeCode" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "lastConsumptionReport" : {
            "type" : "number",
            "format" : "double"
          },
          "username" : {
            "type" : "string"
          }
        },
        "example" : {
          "lastConsumptionReport" : 0.8008281904610115,
          "registrationNumber" : "registrationNumber",
          "name" : "name",
          "typeCode" : "typeCode",
          "username" : "username"
        }
      },
...
      "VehicleInfo" : {
        "type" : "object",
        "properties" : {
          "username" : {
            "type" : "string",
            "description" : "Username of owner of the vehicle."
          },
          "vehicleInfo" : {
            "$ref" : "#/components/schemas/Vehicle"
          }
        }
      },
      "UpdateConsumption" : {
        "type" : "object",
        "properties" : {
          "consumption" : {
            "type" : "string",
            "description" : "consumption to be update."
          }
        }
      }
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}